backend/api/requirements.txt
flask==3.1.0
gunicorn==21.2.0
pika==1.3.2
pymongo==4.9.2
redis==5.2.0
python-redis-lock==4.0.0

# Google OAuth
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1